server: refactor native_lib/rust_imp
