+++ wayland-server/src/event_loop.rs
@@ -1,24 +1,13 @@
-use std::cell::RefCell;
 use std::io::{Error as IoError, Result as IoResult};
-use std::os::raw::{c_int, c_void};
 use std::os::unix::io::RawFd;
 use std::rc::Rc;
 use std::sync::atomic;
 use std::sync::Arc;
 
-use wayland_commons::{downcast_impl, Implementation};
+use Implementation;
 
-use display::DisplayInner;
+use imp::EventLoopInner;
 use sources::{FdEvent, FdInterest, IdleSource, SignalEvent, Source, TimerEvent};
-
-#[cfg(feature = "native_lib")]
-use wayland_sys::server::*;
-
-pub(crate) struct EventLoopInner {
-    #[cfg(feature = "native_lib")]
-    wlevl: *mut wl_event_loop,
-    pub(crate) inner: Option<Rc<DisplayInner>>,
-}
 
 /// An event loop
 ///
@@ -69,30 +58,15 @@ impl LoopSignal {
 impl EventLoop {
     /// Create a new event loop
     pub fn new() -> EventLoop {
-        #[cfg(not(feature = "native_lib"))]
-        {
-            unimplemented!()
-        }
-        #[cfg(feature = "native_lib")]
-        {
-            let ptr = unsafe { ffi_dispatch!(WAYLAND_SERVER_HANDLE, wl_event_loop_create,) };
-            EventLoop {
-                inner: Rc::new(EventLoopInner {
-                    wlevl: ptr,
-                    inner: None,
-                }),
-                stop_signal: Arc::new(atomic::AtomicBool::new(false)),
-            }
+        EventLoop {
+            inner: Rc::new(EventLoopInner::new()),
+            stop_signal: Arc::new(atomic::AtomicBool::new(false)),
         }
     }
 
-    #[cfg(feature = "native_lib")]
-    pub(crate) unsafe fn display_new(disp_inner: Rc<DisplayInner>, ptr: *mut wl_event_loop) -> EventLoop {
+    pub(crate) fn make(inner: EventLoopInner) -> EventLoop {
         EventLoop {
-            inner: Rc::new(EventLoopInner {
-                wlevl: ptr,
-                inner: Some(disp_inner),
-            }),
+            inner: Rc::new(inner),
             stop_signal: Arc::new(atomic::AtomicBool::new(false)),
         }
     }
@@ -118,32 +92,7 @@ impl EventLoop {
     ///
     /// Returns the number of requests dispatched or an error.
     pub fn dispatch(&mut self, timeout: Option<u32>) -> IoResult<u32> {
-        #[cfg(not(feature = "native_lib"))]
-        {
-            unimplemented!()
-        }
-        #[cfg(feature = "native_lib")]
-        {
-            use std::i32;
-            let timeout = match timeout {
-                None => -1,
-                Some(v) if v >= (i32::MAX as u32) => i32::MAX,
-                Some(v) => (v as i32),
-            };
-            let ret = unsafe {
-                ffi_dispatch!(
-                    WAYLAND_SERVER_HANDLE,
-                    wl_event_loop_dispatch,
-                    self.inner.wlevl,
-                    timeout
-                )
-            };
-            if ret >= 0 {
-                Ok(ret as u32)
-            } else {
-                Err(IoError::last_os_error())
-            }
-        }
+        self.inner.dispatch(timeout)
     }
 
     /// Runs the event loop
@@ -162,16 +111,9 @@ impl EventLoop {
     /// set a timeout.
     pub fn run(&mut self) -> IoResult<()> {
         self.stop_signal.store(false, atomic::Ordering::Release);
-        #[cfg(not(feature = "native_lib"))]
-        {
-            unimplemented!()
-        }
-        #[cfg(feature = "native_lib")]
         loop {
-            if let Some(ref display_inner) = self.inner.inner {
-                unsafe { ffi_dispatch!(WAYLAND_SERVER_HANDLE, wl_display_flush_clients, display_inner.ptr) };
-            }
-            self.dispatch(None)?;
+            self.inner.flush_clients_if_display();
+            self.inner.dispatch(None)?;
             if self.stop_signal.load(atomic::Ordering::Acquire) {
                 return Ok(());
             }
@@ -195,33 +137,9 @@ impl LoopToken {
     where
         Impl: Implementation<(), FdEvent> + 'static,
     {
-        #[cfg(not(feature = "native_lib"))]
-        {
-            unimplemented!()
-        }
-        #[cfg(feature = "native_lib")]
-        {
-            let data = Box::new(Box::new(implementation) as Box<Implementation<(), FdEvent>>);
-            let ret = unsafe {
-                ffi_dispatch!(
-                    WAYLAND_SERVER_HANDLE,
-                    wl_event_loop_add_fd,
-                    self.inner.wlevl,
-                    fd,
-                    interest.bits(),
-                    ::sources::event_source_fd_dispatcher,
-                    &*data as *const _ as *mut c_void
-                )
-            };
-            if ret.is_null() {
-                Err((
-                    IoError::last_os_error(),
-                    *(downcast_impl(*data).map_err(|_| ()).unwrap()),
-                ))
-            } else {
-                Ok(Source::make(ret, data))
-            }
-        }
+        self.inner
+            .add_fd_event_source(fd, interest, implementation)
+            .map(Source::make)
     }
 
     /// Add a timer event source to this event loop
@@ -237,31 +155,9 @@ impl LoopToken {
     where
         Impl: Implementation<(), TimerEvent> + 'static,
     {
-        #[cfg(not(feature = "native_lib"))]
-        {
-            unimplemented!()
-        }
-        #[cfg(feature = "native_lib")]
-        {
-            let data = Box::new(Box::new(implementation) as Box<Implementation<(), TimerEvent>>);
-            let ret = unsafe {
-                ffi_dispatch!(
-                    WAYLAND_SERVER_HANDLE,
-                    wl_event_loop_add_timer,
-                    self.inner.wlevl,
-                    ::sources::event_source_timer_dispatcher,
-                    &*data as *const _ as *mut c_void
-                )
-            };
-            if ret.is_null() {
-                Err((
-                    IoError::last_os_error(),
-                    *(downcast_impl(*data).map_err(|_| ()).unwrap()),
-                ))
-            } else {
-                Ok(Source::make(ret, data))
-            }
-        }
+        self.inner
+            .add_timer_event_source(implementation)
+            .map(Source::make)
     }
 
     /// Add a signal event source to this event loop
@@ -278,32 +174,9 @@ impl LoopToken {
     where
         Impl: Implementation<(), SignalEvent> + 'static,
     {
-        #[cfg(not(feature = "native_lib"))]
-        {
-            unimplemented!()
-        }
-        #[cfg(feature = "native_lib")]
-        {
-            let data = Box::new(Box::new(implementation) as Box<Implementation<(), SignalEvent>>);
-            let ret = unsafe {
-                ffi_dispatch!(
-                    WAYLAND_SERVER_HANDLE,
-                    wl_event_loop_add_signal,
-                    self.inner.wlevl,
-                    signal as c_int,
-                    ::sources::event_source_signal_dispatcher,
-                    &*data as *const _ as *mut c_void
-                )
-            };
-            if ret.is_null() {
-                Err((
-                    IoError::last_os_error(),
-                    *(downcast_impl(*data).map_err(|_| ()).unwrap()),
-                ))
-            } else {
-                Ok(Source::make(ret, data))
-            }
-        }
+        self.inner
+            .add_signal_event_source(signal, implementation)
+            .map(Source::make)
     }
 
     /// Add an idle event source to this event loop
@@ -320,50 +193,6 @@ impl LoopToken {
     where
         Impl: Implementation<(), ()> + 'static,
     {
-        #[cfg(not(feature = "native_lib"))]
-        {
-            unimplemented!()
-        }
-        #[cfg(feature = "native_lib")]
-        {
-            let data = Rc::new(RefCell::new((
-                Box::new(implementation) as Box<Implementation<(), ()>>,
-                false,
-            )));
-            let ret = unsafe {
-                ffi_dispatch!(
-                    WAYLAND_SERVER_HANDLE,
-                    wl_event_loop_add_idle,
-                    self.inner.wlevl,
-                    ::sources::event_source_idle_dispatcher,
-                    Rc::into_raw(data.clone()) as *mut _
-                )
-            };
-            IdleSource::make(ret, data)
-        }
-    }
-
-    /// Checks whether given resource is indeed linked to this
-    /// event loop
-    #[cfg(feature = "native_lib")]
-    pub(crate) unsafe fn matches(&self, resource_ptr: *mut wl_resource) -> bool {
-        let client_ptr = ffi_dispatch!(WAYLAND_SERVER_HANDLE, wl_resource_get_client, resource_ptr);
-        let display_ptr = ffi_dispatch!(WAYLAND_SERVER_HANDLE, wl_client_get_display, client_ptr);
-        let event_loop_ptr = ffi_dispatch!(WAYLAND_SERVER_HANDLE, wl_display_get_event_loop, display_ptr);
-        return event_loop_ptr == self.inner.wlevl;
-    }
-}
-
-impl Drop for EventLoopInner {
-    fn drop(&mut self) {
-        #[cfg(feature = "native_lib")]
-        {
-            if self.inner.is_none() {
-                // only destroy the event_loop if it's not the one from the display
-                unsafe {
-                    ffi_dispatch!(WAYLAND_SERVER_HANDLE, wl_event_loop_destroy, self.wlevl);
-                }
-            }
-        }
+        IdleSource::make(self.inner.add_idle_event_source(implementation))
     }
 }
