+++ wayland-server/src/native_lib/mod.rs
@@ -1,0 +1,13 @@
+mod client;
+mod display;
+mod event_loop;
+mod globals;
+mod resource;
+mod source;
+
+pub(crate) use self::client::ClientInner;
+pub(crate) use self::display::DisplayInner;
+pub(crate) use self::event_loop::EventLoopInner;
+pub(crate) use self::globals::GlobalInner;
+pub(crate) use self::resource::{NewResourceInner, ResourceInner};
+pub(crate) use self::source::{IdleSourceInner, SourceInner};
